refactor: 로또 구매가 가능한지에 대한 검증 부분 변경
- 기존 Money에서 로또 1장당 금액을 입력받아 검증하는 부분을 LottoStore에서 담당하도록 변경
- Money는 단순히 전체금액 / 1장당 금액에 대한 결과만을 return하는 divide 기능만을 가지도록 기능 변경
